back-end/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-picklefield==3.3
djangorestframework==3.16.0
drf-spectacular==0.28.0
dwca-writer==0.2.0
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
minio==7.2.15
numpy==2.2.6
openai==1.84.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.3.1
rpds-py==0.25.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.4.0
utm==0.8.1
xlrd==2.0.1
xmltodict==0.14.2
django-allauth==0.60.1
